Add Discounts and Discounts Manager
Previously, there was no discounts database or discounts manager for
applying discounts to products dynamically. Now, there are the Discounts
class which holds the logic for specific discounts like two-for-one and
etc and the Discounts Manager which acts like a database. The way it is structured allows for other products to be added in the hash so that discounts can be dynamically applied. The
basket class has been refactored to accommodate those changes as well.
The previous discounts logic has been moved to it's respective class -
discounts.